+++ CHANGELOG.md
@@ -168,3 +168,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 -test game: casos límite y paso automático de turno
+
+#### [0.3.2] - 2024-09-26
+
+### Added
+-test player: pruebas iniciales de creación y atributos de Player
+
+#### [0.3.3] - 2024-09-26
+
+### Added
+-test player: gestión de fichas en barra y fichas sacadas
+
+#### [0.3.4] - 2024-09-26
+
+### Added
+-test player: independencia entre jugadores, representación y casos límite
+
+#### [0.3.5] - 2024-09-26
+
+### Added
+-test player: casos límite de fichas sacadas y en barra